+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- added half- and full-page scrolling (#22)
+
 ### Fixed
 - `init` command ensures directory of database file exists
 - remove window artefacts after `help` menu is closed (#20)